Event listeners are now just actions
C# delegates are very strict, won't allow a delegate that accepts an
event to be cast to a delegate that accepts a subclass of an event. To
get around this, events just now accept actions and the user can fetch
the most recent event using Event.activeEvent if need be.